requirements.txt
# Zariya B2B Marketplace — Python dependencies
# Optimised for Google Cloud Run (minimal, pinned for reproducibility)

# ── Core framework ────────────────────────────────────────────────────────────
streamlit>=1.35.0,<2.0.0

# ── OAuth (required by st.login) ─────────────────────────────────────────────
Authlib>=1.3.2

# ── Gemini AI (new unified SDK) ───────────────────────────────────────────────
google-genai>=1.0.0

# ── Image handling (logo / assets) ────────────────────────────────────────────
Pillow>=10.3.0

# ── Google Cloud Storage (stub implementation; activate for production) ───────
# google-cloud-storage>=2.16.0
